feat(LHUB019): Reorganized constants packages.
