+++ src/main/java/com/godev/linkhubservice/domain/constants/DatabaseValuesConstants.java
@@ -5,9 +5,9 @@ public class DatabaseValuesConstants {
     private DatabaseValuesConstants(){
     }
 
-    public static final String  DEFAULT_PAGE_PHOTO = "default.png";
-    public static final String  DEFAULT_PAGE_FONT_COLOR = "#212121";
+    public static final String DEFAULT_PAGE_PHOTO = "default.png";
+    public static final String DEFAULT_PAGE_FONT_COLOR = "#212121";
     public static final String DEFAULT_PAGE_BACKGROUND_TYPE_COLOR = "COLOR";
-    public static final String  PAGE_BACKGROUND_TYPE_IMAGE = "IMAGE";
-    public static final String  DEFAULT_PAGE_BACKGROUND_VALUE = "#F4F4F4";
+    public static final String DEFAULT_PAGE_BACKGROUND_VALUE = "#F4F4F4";
+    public static final String PAGE_BACKGROUND_TYPE_IMAGE = "IMAGE";
 }
